server/requirements.txt
numpy
pandas
scipy
scikit-learn
torch
joblib
paho-mqtt
pyserial
matplotlib
shap
